requirements.txt
beautifulsoup4==4.10.0
blis==0.7.5
Bottleneck==1.3.2
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
cycler==0.11.0
cymem==2.0.6
fastai==1.0.58
fastprogress==1.0.0
fonttools==4.28.5
idna==3.3
imageio==2.13.5
imutils==0.5.4
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
langcodes==3.3.0
MarkupSafe==2.0.1
matplotlib==3.5.1
murmurhash==1.0.6
networkx==2.6.3
numexpr==2.8.1
numpy==1.22.0
nvidia-ml-py3==7.352.0
opencv-python==4.5.5.62
packaging==21.3
pandas==1.3.5
pathy==0.6.1
Pillow==9.0.0
preshed==3.0.6
pydantic==1.8.2
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
requests==2.27.1
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
sklearn==0.0
smart-open==5.2.1
soupsieve==2.3.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
thinc==8.0.13
threadpoolctl==3.0.0
tifffile==2021.11.2
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
typer==0.4.0
typing_extensions==4.0.1
urllib3==1.26.8
wasabi==0.9.0
